# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Coffee Land Restaurant ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [reports]
      summary: Admin dashboard — today's KPIs + trends
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Sales report by period
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
          required: true
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/top-items:
    get:
      operationId: getTopItems
      tags: [reports]
      summary: Best-selling menu items
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Top items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopItem"
  /reports/payment-breakdown:
    get:
      operationId: getPaymentBreakdown
      tags: [reports]
      summary: Payment method breakdown
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Payment breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentBreakdown"
  /reports/staff-performance:
    get:
      operationId: getStaffPerformance
      tags: [reports]
      summary: Staff performance metrics
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Staff metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffPerformance"
  /reports/hourly-sales:
    get:
      operationId: getHourlySales
      tags: [reports]
      summary: Hourly sales breakdown for today
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Hourly sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlySales"

  # ─── SETTINGS ──────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get restaurant settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSettings"

    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update restaurant settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSettings"
  /settings/branches:
    get:
      operationId: listBranches
      tags: [settings]
      summary: List branches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"

    post:
      operationId: createBranch
      tags: [settings]
      summary: Create a branch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"

# ─── COMPONENTS ───────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, orderType, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        orderType:
          type: string
          enum: [dine_in, takeaway, delivery]
        status:
          type: string
          enum: [pending, preparing, ready, served, completed, cancelled]
        tableId:
          type: ["integer", "null"]
        tableLabel:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        totalAmount:
          type: number
        discountAmount:
          type: number
        taxAmount:
          type: number
        notes:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DashboardSummary:
      type: object
      required: [todayRevenue, todayOrders, activeOrders, lowStockCount, topItemsToday, recentOrders, ordersByStatus, revenueThisWeek]
      properties:
        todayRevenue:
          type: number
        todayOrders:
          type: integer
        activeOrders:
          type: integer
        lowStockCount:
          type: integer
        avgOrderValue:
          type: number
        topItemsToday:
          type: array
          items:
            $ref: "#/components/schemas/TopItem"
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        ordersByStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        revenueThisWeek:
          type: array
          items:
            $ref: "#/components/schemas/DailyRevenue"
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
    DailyRevenue:
      type: object
      required: [date, revenue, orders]
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    SalesReport:
      type: object
      required: [totalRevenue, totalOrders, avgOrderValue, data, topCategories]
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        avgOrderValue:
          type: number
        data:
          type: array
          items:
            $ref: "#/components/schemas/DailyRevenue"
        topCategories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryRevenue"
    CategoryRevenue:
      type: object
      required: [categoryName, revenue, orders]
      properties:
        categoryName:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    TopItem:
      type: object
      required: [menuItemId, nameEn, nameAm, totalSold, revenue]
      properties:
        menuItemId:
          type: integer
        nameEn:
          type: string
        nameAm:
          type: string
        totalSold:
          type: integer
        revenue:
          type: number
    PaymentBreakdown:
      type: object
      required: [total, breakdown]
      properties:
        total:
          type: number
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMethodStat"
    PaymentMethodStat:
      type: object
      required: [method, count, amount, percentage]
      properties:
        method:
          type: string
        count:
          type: integer
        amount:
          type: number
        percentage:
          type: number
    StaffPerformance:
      type: object
      required: [employeeId, employeeName, role, ordersHandled, totalRevenue, avgOrderTime]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        role:
          type: string
        ordersHandled:
          type: integer
        totalRevenue:
          type: number
        avgOrderTime:
          type: number
    HourlySales:
      type: object
      required: [hour, revenue, orders]
      properties:
        hour:
          type: integer
        revenue:
          type: number
        orders:
          type: integer

    # ── SETTINGS ──────────────────────────────────────────────────────────────
    RestaurantSettings:
      type: object
      required: [id, name, branchId]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAm:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        vatRate:
          type: number
        loyaltyPointsPerEtb:
          type: number
        receiptFooterText:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        branchId:
          type: integer
        updatedAt:
          type: string
          format: date-time
    RestaurantSettingsUpdate:
      type: object
      properties:
        name:
          type: string
        nameAm:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        vatRate:
          type: number
        loyaltyPointsPerEtb:
          type: number
        receiptFooterText:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
    Branch:
      type: object
      required: [id, name, address, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BranchInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean